requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-paypal==2.0
django-versatileimagefield==2.2
idna==3.4
Pillow==9.3.0
psycopg2-binary==2.9.5
python-magic==0.4.27
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.12
